rename fused elementwise
